==> landmarks/1.select_model.ipynb <==
import pandas as pd

df_input = pd.read_csv("../data/face/df_input.csv")

features = pd.read_csv("../data/face/features.csv")

df_input.columns

features.columns

df_input.drop(['Unnamed: 0', 'file'], axis=1, inplace=True)

df_input

from pycaret.regression import *

face_inputs = setup(data = df_input, target='Rating', session_id=123)

compare_models()
